+++ MyLayoutDemo/AllTest12ViewController.h
@@ -1,0 +1,16 @@
+//
+//  AllTest11ViewController.h
+//  MyLayoutDemo
+//
+//  Created by oubaiquan on 2018/8/1.
+//  Copyright © 2018年 YoungSoft. All rights reserved.
+//
+
+#import <UIKit/UIKit.h>
+
+/**
+*❁6.MyLayout & AutoLayout
+*/
+@interface AllTest12ViewController : UIViewController
+
+@end