+++ MyLayoutDemo/AllTest1TableViewCellForAutoLayout.h
@@ -1,8 +1,8 @@
 //
-//  AllTest1TableViewCell.h
+//  AllTest1TableViewCellForAutoLayout.h
 //  MyLayout
 //
-//  Created by apple on 16/5/25.
+//  Created by apple on 19/5/16.
 //  Copyright © 2016年 YoungSoft. All rights reserved.
 //
 
@@ -11,10 +11,10 @@
 #import "AllTestDataModel.h"
 
 /**
- * 动态高度UITableViewCell
- * 本例子是将布局视图的高度自适应能力和在不使用AutoLayout的情况下实现UITableViewCell高度自适应的能力
+ * 动态高度UITableViewCellForAutoLayout
+ * 本例子是将布局视图的高度自适应能力和使用Autolayout的情况下实现UITableViewCell高度自适应的能力
  */
-@interface AllTest1TableViewCell : UITableViewCell
+@interface AllTest1TableViewCellForAutoLayout : UITableViewCell
 
 //对于需要动态评估高度的UITableViewCell来说可以把布局视图暴露出来。用于高度评估和边界线处理。以及事件处理的设置。
 @property(nonatomic, strong, readonly) MyBaseLayout *rootLayout;